State Persistence
- Persisted state into local storage.
- abstracted components in Movie page.
